requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Object Storage (MinIO/S3)
minio==7.2.0
boto3==1.34.1

# YAML configuration
PyYAML==6.0.1
jsonschema==4.20.0

# HTTP clients
requests==2.31.0
urllib3==2.1.0

# Report generation
jinja2==3.1.2
weasyprint==60.1  # For PDF generation
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Notifications
slack-sdk==3.26.1
python-gitlab==4.2.0
PyGithub==2.1.1

# Email
secure-smtplib==0.1.1

# Logging and monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Security
cryptography==41.0.7
pyjwt==2.8.0

# Task scheduling
celery==5.3.4
redis==5.0.1

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
tabulate==0.9.0
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
responses==0.24.1
faker==20.1.0

# Code quality
flake8==6.1.0
pylint==3.0.3
black==23.12.1
isort==5.13.2
mypy==1.7.1

# Development
ipython==8.18.1
ipdb==0.13.13
watchdog==3.0.0
